--- src/OHMFRAXBondV1.ts
@@ -1,18 +1,0 @@
-import { DepositCall } from "../generated/OHMFRAXBondV1/OHMFRAXBondV1";
-import { createDailyBondRecord } from "./bonds/DailyBond";
-import { OHMFRAXLPBOND_TOKEN, PAIR_UNISWAP_V2_OHM_FRAX } from "./utils/Constants";
-import { toDecimal } from "./utils/Decimals";
-import { getOhmUSDPairValue } from "./utils/LiquidityUniswapV2";
-import { loadOrCreateToken } from "./utils/Tokens";
-
-export function handleDeposit(call: DepositCall): void {
-  const token = loadOrCreateToken(OHMFRAXLPBOND_TOKEN);
-  const amount = toDecimal(call.inputs._amount, 18);
-
-  createDailyBondRecord(
-    call.block.timestamp,
-    token,
-    amount,
-    getOhmUSDPairValue(call.inputs._amount, PAIR_UNISWAP_V2_OHM_FRAX, call.block.number),
-  );
-}
