+++ src/protocolMetrics/Rebase.ts
@@ -1,0 +1,59 @@
+import { Address, BigDecimal, BigInt } from "@graphprotocol/graph-ts";
+import { log } from "matchstick-as";
+
+import { OlympusStakingV1 } from "../../generated/ProtocolMetrics/OlympusStakingV1";
+import { OlympusStakingV2 } from "../../generated/ProtocolMetrics/OlympusStakingV2";
+import { OlympusStakingV3 } from "../../generated/ProtocolMetrics/OlympusStakingV3";
+import {
+  STAKING_CONTRACT_V1,
+  STAKING_CONTRACT_V2,
+  STAKING_CONTRACT_V2_BLOCK,
+  STAKING_CONTRACT_V3,
+  STAKING_CONTRACT_V3_BLOCK,
+} from "../utils/Constants";
+import { toDecimal } from "../utils/Decimals";
+
+export function getNextOHMRebase(blockNumber: BigInt): BigDecimal {
+  let next_distribution = BigDecimal.fromString("0");
+
+  const staking_contract_v1 = OlympusStakingV1.bind(Address.fromString(STAKING_CONTRACT_V1));
+  const response = staking_contract_v1.try_ohmToDistributeNextEpoch();
+  if (response.reverted == false) {
+    next_distribution = toDecimal(response.value, 9);
+    log.debug("next_distribution v1 {}", [next_distribution.toString()]);
+  } else {
+    log.debug("reverted staking_contract_v1", []);
+  }
+
+  if (blockNumber.gt(BigInt.fromString(STAKING_CONTRACT_V2_BLOCK))) {
+    const staking_contract_v2 = OlympusStakingV2.bind(Address.fromString(STAKING_CONTRACT_V2));
+    const distribution_v2 = toDecimal(staking_contract_v2.epoch().value3, 9);
+    log.debug("next_distribution v2 {}", [distribution_v2.toString()]);
+    next_distribution = next_distribution.plus(distribution_v2);
+  }
+
+  if (blockNumber.gt(BigInt.fromString(STAKING_CONTRACT_V3_BLOCK))) {
+    const staking_contract_v3 = OlympusStakingV3.bind(Address.fromString(STAKING_CONTRACT_V3));
+    const distribution_v3 = toDecimal(staking_contract_v3.epoch().value3, 9);
+    log.debug("next_distribution v3 {}", [distribution_v3.toString()]);
+    next_distribution = next_distribution.plus(distribution_v3);
+  }
+
+  log.debug("next_distribution total {}", [next_distribution.toString()]);
+
+  return next_distribution;
+}
+
+export function getAPY_Rebase(sOHM: BigDecimal, distributedOHM: BigDecimal): BigDecimal[] {
+  const nextEpochRebase = distributedOHM.div(sOHM).times(BigDecimal.fromString("100"));
+
+  const nextEpochRebase_number = parseFloat(nextEpochRebase.toString());
+  const currentAPY = (Math.pow(nextEpochRebase_number / 100 + 1, 365 * 3) - 1) * 100;
+
+  const currentAPYdecimal = BigDecimal.fromString(currentAPY.toString());
+
+  log.debug("next_rebase {}", [nextEpochRebase.toString()]);
+  log.debug("current_apy total {}", [currentAPYdecimal.toString()]);
+
+  return [currentAPYdecimal, nextEpochRebase];
+}
