+++ src/sOlympus/sOlympusERC20V2.ts
@@ -1,12 +1,12 @@
 import { Address, BigInt } from "@graphprotocol/graph-ts";
 
-import { Rebase } from "../generated/schema";
-import { OlympusERC20 } from "../generated/sOlympusERC20V2/OlympusERC20";
-import { RebaseCall } from "../generated/sOlympusERC20V2/sOlympusERC20V2";
-import { ERC20_OHM_V1, STAKING_CONTRACT_V2 } from "./utils/Constants";
-import { createDailyStakingReward } from "./utils/DailyStakingReward";
-import { toDecimal } from "./utils/Decimals";
-import { getBaseOhmUsdRate } from "./utils/Price";
+import { Rebase } from "../../generated/schema";
+import { OlympusERC20 } from "../../generated/sOlympusERC20V2/OlympusERC20";
+import { RebaseCall } from "../../generated/sOlympusERC20V2/sOlympusERC20V2";
+import { ERC20_OHM_V1, STAKING_CONTRACT_V2 } from "../utils/Constants";
+import { createDailyStakingReward } from "../utils/DailyStakingReward";
+import { toDecimal } from "../utils/Decimals";
+import { getBaseOhmUsdRate } from "../utils/Price";
 
 export function rebaseFunction(call: RebaseCall): void {
   let rebase = Rebase.load(call.block.timestamp.toString());
