+++ src/sOlympus/sOlympusERC20V3.ts
@@ -1,12 +1,12 @@
 import { Address, BigInt } from "@graphprotocol/graph-ts";
 
-import { Rebase } from "../generated/schema";
-import { OlympusERC20 } from "../generated/sOlympusERC20V3/OlympusERC20";
-import { RebaseCall } from "../generated/sOlympusERC20V3/sOlympusERC20V3";
-import { ERC20_OHM_V2, STAKING_CONTRACT_V3 } from "./utils/Constants";
-import { createDailyStakingReward } from "./utils/DailyStakingReward";
-import { toDecimal } from "./utils/Decimals";
-import { getBaseOhmUsdRate } from "./utils/Price";
+import { Rebase } from "../../generated/schema";
+import { OlympusERC20 } from "../../generated/sOlympusERC20V3/OlympusERC20";
+import { RebaseCall } from "../../generated/sOlympusERC20V3/sOlympusERC20V3";
+import { ERC20_OHM_V2, STAKING_CONTRACT_V3 } from "../utils/Constants";
+import { createDailyStakingReward } from "../utils/DailyStakingReward";
+import { toDecimal } from "../utils/Decimals";
+import { getBaseOhmUsdRate } from "../utils/Price";
 
 export function rebaseFunction(call: RebaseCall): void {
   let rebase = Rebase.load(call.block.timestamp.toString());
